Keep editors open instead of closing them when loading fails on startup (fix #110062)
